requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
python-json-logger==3.2.1
sqlparse==0.5.3
